obsidian-rag-mcp/requirements.txt
mcp
chromadb
sentence-transformers
python-frontmatter
pyyaml
watchdog
tiktoken>=0.5.0
